+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2021-12-17
+
+### Added
+- Flash loan endpoint which allows a tech savvy user to borrow any amount of
+  liquidity. Design inspired by
+  [Solana labs implementation](https://github.com/solana-labs/solana-program-library/blob/a6a1ce4290ad683e13b9ef1c52d6f32a080a996f/token-lending/program/src/processor.rs#L1521).
+- Endpoint to toggle flash loans feature. Flash loans are by default disabled.
+
 ## [0.1.1] - 2021-12-14
 
 ### Fixed